Make all links open in new tabs (#74)
* Make all links open in new tabs

This makes it so that all the links in the premid troubleshooter open in a new tab. You may not want all of these to have this behavior so please let me know and I'll remove the ones you don't want but I believe most of these are necessary so that the user doesn't lose progress when clicking any links.

* Changed translation commit message

Since your only file is called en.json, I feel like it would make more sense to remove the name of the file and only state the language that was changed. If more files are added in the future, you can always revert this.